Move _find_source to pathfinding
